+++ src/main/java/uk/co/transactis/Triangle.java
@@ -12,7 +12,7 @@ public class Triangle implements Shape {
     private float TRIANGLE_WIDTH=10f;
     private float TRIANGLE_HEIGHT=10f;
 
-    public float getArea() {
+    public double getArea() {
         return 0.5f * TRIANGLE_HEIGHT * TRIANGLE_WIDTH;
     }
 }
